fix: resolve SonarCloud quality gate issues and lowercase API routes
- Add SuppressMessage(S1144) on EF entity Id properties (6 issues)
- Refactor 7 service methods to reduce cognitive complexity below 15 (S3776)
- Fix null dereferences in MiningMachineQueryService/SlotQueryService (CS8602)
- Make worker helper methods static (S2325/CA1822)
- Remove unnecessary decimal cast in MiningEngine (S1905)
- Split TokensController: move candle/mining-rule to dedicated controllers (S6960)
- Reorder TradingContextMapper overloads to be adjacent (S4136)
- Extract static readonly arrays in tests (CA1861)
- Fix missing XML param tag in MainDtos (CS1573)
- Add XML comments to AccessControlService/AccessSetting (CS1591)
- Normalize API routes to lowercase: api/v1/tokens

+++ ArkWallet/Domain/Engines/MiningEngine.cs
@@ -85,7 +85,7 @@ internal class MiningEngine
         var elapsed = (now - start.Value).TotalMinutes;
         var total = (end.Value - start.Value).TotalMinutes;
         var percent = (decimal)(elapsed / total) * 100m;
-        return Math.Clamp((decimal)percent, 0m, 100m);
+        return Math.Clamp(percent, 0m, 100m);
     }
 
     /// <summary>
